requirements.txt
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.5.0
attrs==24.2.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dill==0.3.9
easyocr==1.7.2
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.5
idna==3.7
imageio==2.36.1
inference-sdk==0.29.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.17.0
networkx==3.4.2
ninja==1.11.1.2
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyclipper==1.3.0.post6
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.49
rpds-py==0.22.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
smmap==5.0.1
streamlit==1.41.0
supervision==0.22.0
sympy==1.13.1
tenacity==9.0.0
thop==0.1.1.post2209072238
tifffile==2024.9.20
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1+cpu
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.40
ultralytics-thop==2.0.13
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.18.3
